Split up the types, traits and deserializers
