--- src/trait_implementations/mod.rs
@@ -1,6 +1,0 @@
-pub mod dbi;
-pub mod login;
-mod ringing;
-mod substitution;
-pub mod timeline;
-mod timetable;
