+++ src/traits/dbi.rs
@@ -1,12 +1,31 @@
 use crate::{
-    edupage::{Edupage, EdupageError},
-    edupage_traits::DBI,
+    edupage::EdupageError, edupage::Edupage, types::dbi::*, types::person::*,
+    types::timeline::*, types::timetable::Timetable as EduTimetable,
+    types::RingingTime,
 };
+use chrono::{NaiveDate, NaiveDateTime};
+use serde::{Deserialize, Serialize};
+
+
+pub trait DBI {
+    fn get_teachers(&self) -> Result<Vec<Teacher>, EdupageError>;
+    fn get_teacher_by_id(&self, id: i64) -> Result<Option<Teacher>, EdupageError>;
+
+    fn get_students(&self) -> Result<Vec<Student>, EdupageError>;
+    fn get_student_by_id(&self, id: i64) -> Result<Option<Student>, EdupageError>;
+
+    fn get_subjects(&self) -> Result<Vec<DBIBase>, EdupageError>;
+    fn get_subject_by_id(&self, id: i64) -> Result<Option<DBIBase>, EdupageError>;
+
+    fn get_classrooms(&self) -> Result<Vec<DBIBase>, EdupageError>;
+    fn get_classroom_by_id(&self, id: i64) -> Result<Option<DBIBase>, EdupageError>;
+}
+
 
 impl DBI for Edupage {
     fn get_teachers(
         &self,
-    ) -> Result<Vec<crate::edupage_types::person::Teacher>, crate::edupage::EdupageError> {
+    ) -> Result<Vec<crate::types::person::Teacher>, crate::edupage::EdupageError> {
         if !self.is_logged_in {
             return Err(EdupageError::NotLoggedIn);
         }
@@ -19,7 +38,7 @@ impl DBI for Edupage {
     fn get_teacher_by_id(
         &self,
         id: i64,
-    ) -> Result<Option<crate::edupage_types::person::Teacher>, crate::edupage::EdupageError> {
+    ) -> Result<Option<crate::types::person::Teacher>, crate::edupage::EdupageError> {
         if !self.is_logged_in {
             return Err(EdupageError::NotLoggedIn);
         }
@@ -37,7 +56,7 @@ impl DBI for Edupage {
 
     fn get_students(
         &self,
-    ) -> Result<Vec<crate::edupage_types::person::Student>, crate::edupage::EdupageError> {
+    ) -> Result<Vec<crate::types::person::Student>, crate::edupage::EdupageError> {
         if !self.is_logged_in {
             return Err(EdupageError::NotLoggedIn);
         }
@@ -50,7 +69,7 @@ impl DBI for Edupage {
     fn get_student_by_id(
         &self,
         id: i64,
-    ) -> Result<Option<crate::edupage_types::person::Student>, crate::edupage::EdupageError> {
+    ) -> Result<Option<crate::types::person::Student>, crate::edupage::EdupageError> {
         if !self.is_logged_in {
             return Err(EdupageError::NotLoggedIn);
         }
@@ -68,7 +87,7 @@ impl DBI for Edupage {
 
     fn get_subjects(
         &self,
-    ) -> Result<Vec<crate::edupage_types::dbi::DBIBase>, crate::edupage::EdupageError> {
+    ) -> Result<Vec<crate::types::dbi::DBIBase>, crate::edupage::EdupageError> {
         if !self.is_logged_in {
             return Err(EdupageError::NotLoggedIn);
         }
@@ -81,7 +100,7 @@ impl DBI for Edupage {
     fn get_subject_by_id(
         &self,
         id: i64,
-    ) -> Result<Option<crate::edupage_types::dbi::DBIBase>, crate::edupage::EdupageError> {
+    ) -> Result<Option<crate::types::dbi::DBIBase>, crate::edupage::EdupageError> {
         if !self.is_logged_in {
             return Err(EdupageError::NotLoggedIn);
         }
@@ -99,7 +118,7 @@ impl DBI for Edupage {
 
     fn get_classrooms(
         &self,
-    ) -> Result<Vec<crate::edupage_types::dbi::DBIBase>, crate::edupage::EdupageError> {
+    ) -> Result<Vec<crate::types::dbi::DBIBase>, crate::edupage::EdupageError> {
         if !self.is_logged_in {
             return Err(EdupageError::NotLoggedIn);
         }
@@ -112,7 +131,7 @@ impl DBI for Edupage {
     fn get_classroom_by_id(
         &self,
         id: i64,
-    ) -> Result<Option<crate::edupage_types::dbi::DBIBase>, crate::edupage::EdupageError> {
+    ) -> Result<Option<crate::types::dbi::DBIBase>, crate::edupage::EdupageError> {
         if !self.is_logged_in {
             return Err(EdupageError::NotLoggedIn);
         }
