+++ src/traits/mod.rs
@@ -1,0 +1,14 @@
+pub mod dbi;
+pub mod login;
+pub mod ringing;
+pub mod substitution;
+pub mod timeline;
+pub mod timetable;
+
+
+pub use dbi::*;
+pub use login::*;
+pub use ringing::*;
+pub use substitution::*;
+pub use timeline::*;
+pub use timetable::*;
