+++ src/types/person.rs
@@ -2,7 +2,7 @@
 use chrono::NaiveDate;
 use serde::{Deserialize, Serialize};
 
-use crate::edupage_deserializers::*;
+use crate::deserializers::*;
 use crate::macro_aliases::*;
 
 use super::dbi::Gender;
