+++ src/types/timetable.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 use chrono::{NaiveDate, NaiveDateTime};
 use serde::{Deserialize, Serialize};
 
-use crate::edupage_deserializers::*;
+use crate::deserializers::*;
 use crate::macro_aliases::*;
 
 use super::dbi::DBIBase;
